added all the tools, a second datadisk, better handling of the node install and terminal setup, added logic to check previous installed status

+++ scripts/atom.sh
@@ -3,11 +3,13 @@
 echo "***** Running atom *****"
 version=1.12.6
 
-#cd /opt
+if [ ! -f /usr/bin/atom ]
+then
 
-echo "*** downloading $version"
-curl -O -L https://github.com/atom/atom/releases/download/v$version/atom.x86_64.rpm
-yum localinstall atom.x86_64.rpm -y
-rm atom.x86_64.rpm
+    echo "*** downloading $version"
+    curl -O -L https://github.com/atom/atom/releases/download/v$version/atom.x86_64.rpm
+    yum localinstall atom.x86_64.rpm -y
+    rm atom.x86_64.rpm
+fi
 
 echo "***** Fin! atom *****"
