+++ scripts/mongotron.sh
@@ -3,13 +3,15 @@
 echo "***** Running mongotron *****"
 version=1.0.0-alpha.5
 
-#cd /opt
+pushd /opt
 
-echo "*** installing mongotron $version ***"
+if [ ! -f /opt/Mongotron-linux-x64/Mongotron ]
+then
+    echo "*** installing mongotron $version ***"
 
-curl -O -L https://github.com/officert/mongotron/releases/download/$version/Mongotron-linux-x64.zip
-unzip Mongotron-linux-x64.zip
-rm Mongotron-linux-x64.zip
+    curl -O -L https://github.com/officert/mongotron/releases/download/$version/Mongotron-linux-x64.zip
+    unzip Mongotron-linux-x64.zip
+    rm Mongotron-linux-x64.zip
 
 cat<<EOF > /usr/share/applications/mongotron.desktop
 [Desktop Entry]
@@ -22,5 +24,9 @@ Exec=/opt/Mongotron-linux-x64/Mongotron
 Terminal=false
 Categories=Development;
 EOF
+
+fi
+
+popd
 
 echo "***** Fin! mongotron *****"
