+++ scripts/webstorm.sh
@@ -2,17 +2,16 @@
 
 echo "***** Running webstorm *****"
 version="WebStorm-2016.3.1"
-#cd /opt
+pushd /opt
 
-#echo "*** installing java 1.7 and libc"
-#yum install java-1.7.0-openjdk -y
-#yum install libc.so.6 -y
+if [ ! -f /usr/share/applications/webstorm.desktop ]
+then 
+    echo "*** installing webstorm $version ***"
+    curl -O -L https://download.jetbrains.com/webstorm/$version.tar.gz
 
-echo "*** installing webstorm $version ***"
-curl -O -L https://download.jetbrains.com/webstorm/$version.tar.gz
-          #https://download-cf.jetbrains.com/webstorm/$version.tar.gz
-tar xfz $version.tar.gz -C /opt/$version
-rm $version.tar.gz
+    tar xfz $version.tar.gz
+    rm $version.tar.gz
+    mv WebStorm-* $version
 
 cat<<EOF > /usr/share/applications/webstorm.desktop
 [Desktop Entry]
@@ -27,5 +26,9 @@ Icon=/opt/$version/bin/webstorm.svg
 Terminal=false
 Categories=Development;
 EOF
+
+fi
+
+popd
 
 echo "***** Fin! webstorm *****"
